% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm2.R
\name{algorithm2}
\alias{algorithm2}
\title{SpatialDecon mixed cell deconvolution algorithm}
\usage{
algorithm2(
  Y,
  X,
  bg = 0,
  weights = NULL,
  resid_thresh = 3,
  lower_thresh = 0.5,
  align_genes = TRUE,
  maxit = 1000
)
}
\arguments{
\item{Y}{p-length expression vector or p * N expression matrix - the actual
(linear-scale) data}

\item{X}{p * K Training matrix.}

\item{bg}{Expected background counts. Provide a scalar to apply to all
data points, or
else a matrix/vector aligning with Y to provide more nuanced expected
 background.}

\item{weights}{The same as the weights argument used by lm}

\item{resid_thresh}{A scalar, sets a threshold on how extreme individual
data points' values
 can be (in log2 units) before getting flagged as outliers and set to NA.}

\item{lower_thresh}{A scalar. Before log2-scale residuals are calculated,
both observed and fitted
 values get thresholded up to this value. Prevents log2-scale residuals from
 becoming extreme in
 points near zero.}

\item{align_genes}{Logical. If TRUE, then Y, X, bg, and wts are row-aligned
by shared genes.}

\item{maxit}{Maximum number of iterations. Default 1000.}
}
\value{
a list:
\itemize{
\item beta: matrix of cell abundance estimates, cells in rows and
observations in columns
\item sigmas: covariance matrices of each observation's beta estimates
\item p: matrix of p-values for H0: beta == 0
\item t: matrix of t-statistics for H0: beta == 0
\item se: matrix of standard errors of beta values
\item resids: a matrix of residuals from the model fit.
(log2(pmax(y, lower_thresh)) - log2(pmax(xb, lower_thresh))).
}
}
\description{
Runs the generic SpatialDecon decon workflow, including:
\enumerate{
\item run deconvolution once
\item remove poorly-fit genes from first round of decon
\item re-run decon with cleaned-up gene set
\item compute p-values
}
}
