% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{andersson_g1}
\alias{andersson_g1}
\title{A Spatial Transcriptome Seurat Object}
\format{
A Seurat Object with a Spatial Assay
}
\usage{
andersson_g1
}
\description{
Andersson, A. et al. Spatial Deconvolution of HER2-positive Breast Tumors 
Reveals Novel Intercellular Relationships. 
http://biorxiv.org/lookup/doi/10.1101/2020.07.14.200600 (2020) doi:10.1101/2020.07.14.200600.
}
\keyword{datasets}
