% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCellScoresToCounts.R
\name{convertCellScoresToCounts}
\alias{convertCellScoresToCounts}
\title{Convert abundance measurements to cell counts}
\usage{
convertCellScoresToCounts(beta, nuclei.counts = NULL, omit.tumor = FALSE)
}
\arguments{
\item{beta}{Matrix of cell abundance scores, with cells in rows and
observations in columns. The
assumption is that this matrix is from well-normalized data.}

\item{nuclei.counts}{Optional. A vector of total nuclei counts. If provided,
the function will
 output not only cells.per.100 but also total cells.}

\item{omit.tumor}{Logical. If FALSE, any rows of beta with "tumor" in their
name will be omitted.}
}
\value{
A list with two elements, each a rescaled version of beta.
cells.per.100 gives estimated
 percents of total, and cell.counts is cells.per.100 * nuclei.counts.
}
\description{
Converts cell abundance scores to cell counts, under the assumption that the
 observation with
 the greatest sum of cell abundance scores is 100% modelled cells.
}
