% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveWeights.R
\name{deriveWeights}
\alias{deriveWeights}
\title{Compute gene weights}
\usage{
deriveWeights(
  norm,
  raw = NULL,
  error.model = "dsp",
  weight.by.TIL.resid.sd = FALSE
)
}
\arguments{
\item{norm}{Matrix of data used in decon. Used to define the gene list and
the
shape of the output "wts" matrix.}

\item{raw}{Matrix of raw data. If provided, used to define technical noise}

\item{error.model}{Which error model to use. Defaults to "dsp"}

\item{weight.by.TIL.resid.sd}{If TRUE, then genes are weighted in part based
on their
 biological variability as estimated by their residual SD from decon
  performed on TCGA.}
}
\value{
A matrix of weights, in the same dimension as norm
}
\description{
Compute gene weights from pre-defined biological noise estimates and from
 raw count/ error-model-defined technical noise estimates
}
