% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_profile_matrix.R
\name{download_profile_matrix}
\alias{download_profile_matrix}
\title{Download a cell profile matrix}
\usage{
download_profile_matrix(species, age_group, matrixname)
}
\arguments{
\item{species}{species of profile matrix}

\item{age_group}{age_group of profile matrix, if fetal mouse please add the 
developmental stage separated with /, i.e. Fetal/E14.5}

\item{matrixname}{name of profile matrix}
}
\value{
A cell profile matrix, suggested cell groups, and paper metadata
}
\description{
Download a cell profile matrix from the online library
}
\details{
Valid matrices can be found on the github site 
\url{https://github.com/Nanostring-Biostats/CellProfileLibrary/tree/NewProfileMatrices}
}
\examples{
download_profile_matrix(species = "Human", age_group = "Adult", matrixname = "Colon_HCA")
head(profile_matrix)
print(cellGroups)
print(metadata)
}
