% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean.resid.sd}
\alias{mean.resid.sd}
\title{Genes' biological variability in immune deconvolution from TCGA.}
\format{
A named vector giving SDs of 1179 genes.
}
\usage{
mean.resid.sd
}
\description{
Genes' biological SDs, as estimated from immune deconvolution from TCGA.
Used to weight genes in spatialdecon.
}
\keyword{datasets}
