% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTumorIntoX.R
\name{mergeTumorIntoX}
\alias{mergeTumorIntoX}
\title{Estimate a tumor-specific profile and merge it with the pre-specified cell profile matrix (X)}
\usage{
mergeTumorIntoX(norm, bg, pure.tumor.ids, X, K = 10)
}
\arguments{
\item{norm}{matrix of normalized data}

\item{bg}{matrix of expected background, on the scale of norm.}

\item{pure.tumor.ids}{Vector identifying columns of norm that are pure tumor. 
Can be indices, logicals or column names.}

\item{X}{The training matrix}

\item{K}{the number of clusters to fit}
}
\value{
an updated X matrix with new columns, "tumor.1", "tumor.2", ...
}
\description{
Given the input of "tumor-only" AOI's, estimates an collection of tumor-specific 
expression profiles and merges them with the immune cell expression training matrix. 
The process: 
\enumerate{
\item log2/normalized data from tumor-only AOIs is clustered with hclust, 
and cutree() is used to define clusters. 
\item 2. Each cluster's geomean profile is merged into the immune cell profile matrix.
}
}
