% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_geomx_dataset}
\alias{mini_geomx_dataset}
\title{Example GeoMx data}
\format{
A list with the following elements:
 \itemize{
 \item normalized: normalized data matrix
 \item raw: raw data matrix
 \item annot: AOI annotation data frane
 }
}
\usage{
mini_geomx_dataset
}
\description{
A miniature GeoMx dataset used by the spatialdecon examples.
}
\keyword{datasets}
