% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_singleCell_dataset}
\alias{mini_singleCell_dataset}
\title{Mini human colon single cell dataset}
\format{
A list with the following elements:
 \itemize{
 \item mtx: sparse count matrix
 \item annots: cell type annotation data frame
 }
}
\usage{
mini_singleCell_dataset
}
\description{
Random 250 cells and most informative genes (CV > 10) between cell types from
Kinchen, J. et al. Structural Remodeling of the Human Colonic 
Mesenchyme in Inflammatory Bowel Disease. Cell 175, 372-386.e17 (2018).
}
\keyword{datasets}
