% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsclc}
\alias{nsclc}
\title{Large example GeoMx data}
\format{
GeoMxSet Object
}
\usage{
nsclc
}
\description{
A GeoMx dataset with dense AOIs gridded over a NSCLC tumor. Each AOI is split into tumor and microenvironment segments.
}
\keyword{datasets}
