% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmergeTumorIntoXGeo.R
\name{runMergeTumorIntoX}
\alias{runMergeTumorIntoX}
\title{Run MergeTumorIntoX}
\usage{
runMergeTumorIntoX(object, ...)
}
\arguments{
\item{object}{An S4 object such as a GeoMxSet object}

\item{...}{Arguments passed to mergeTumorIntoX}
}
\value{
updated X matrix with new columns, "tumor.1", "tumor.2", ...
}
\description{
Runs mergeTumorIntoX from an S4 object
}
\examples{
library(GeomxTools)
datadir <- system.file("extdata", "DSP_NGS_Example_Data", package = "GeomxTools")
demoData <- readRDS(file.path(datadir, "/demoData.rds"))

demoData <- shiftCountsOne(demoData)
target_demoData <- aggregateCounts(demoData)

target_demoData <- normalize(target_demoData, "quant")

data(safeTME)
tumor.ids <- as.logical(sample(x = c("TRUE","FALSE"), size = 88, replace = TRUE))
safeTME.with.tumor <- runMergeTumorIntoX(object = target_demoData,
                                         X = safeTME,
                                         K = 3,
                                         pure_tumor_ids = tumor.ids,
                                         norm_elt = "exprs_norm")
}
