% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runreverseDeconGeo.R
\name{runReverseDecon}
\alias{runReverseDecon}
\title{Run Reversedecon}
\usage{
runReverseDecon(object, ...)
}
\arguments{
\item{object}{An S4 object such as a GeoMxSet object}

\item{...}{Arguments passed to reversedecon}
}
\value{
list containing modeled gene expression's ~ cell scores
}
\description{
Runs reversedecon from an S4 object
}
\examples{
library(GeomxTools)
datadir <- system.file("extdata", "DSP_NGS_Example_Data", package = "GeomxTools")
demoData <- readRDS(file.path(datadir, "/demoData.rds"))

demoData <- shiftCountsOne(demoData)
target_demoData <- aggregateCounts(demoData)

target_demoData <- normalize(target_demoData, "quant")
                
# run basic decon:
res0 <- runspatialdecon(object = target_demoData,
                        norm_elt = "exprs_norm",
                        raw_elt = "exprs")

# run reverse decon:
target_demoData <- runReverseDecon(object = target_demoData,
                                   norm_elt = "exprs_norm",
                                   beta = pData(res0)$beta)

}
