% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{safeTME}
\alias{safeTME}
\title{SafeTME matrix}
\format{
A matrix with 906 genes (rows) and 18 cell types (columns)
}
\usage{
safeTME
}
\description{
A matrix of expression profiles of 906 genes over 18 cell types.
}
\keyword{datasets}
