% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm2.R
\name{tidy_X_and_Y}
\alias{tidy_X_and_Y}
\title{Function to format Y, X inputs for decon}
\usage{
tidy_X_and_Y(X, Y)
}
\arguments{
\item{X}{X matrix}

\item{Y}{Data matrix}
}
\value{
X and Y, both formatted as matrices, with full dimnames and aligned
to each other by dimname
}
\description{
Takes user-supplied X and Y, checks for accuracy, aligns by dimnames, adds
dimnames if missing
}
