% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_geomx_dataset}
\alias{mini_geomx_dataset}
\title{Default colors for the cell types in the safeTME matrix}
\format{
A named vector
}
\usage{
mini_geomx_dataset
}
\description{
A named vector of colors, giving colors for the cell types of the safeTME
 matrix.
}
\keyword{datasets}
