% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsclc}
\alias{nsclc}
\title{Large example GeoMx data}
\format{
A list with the following elements:
 \itemize{
 \item normalized: normalized data matrix
 \item raw: raw data matrix
 \item annot: AOI annotation data frane
 }
}
\usage{
nsclc
}
\description{
A GeoMx dataset with dense AOIs gridded over a NSCLC tumor. Each AOI is split into tumor and microenvironment segments.
}
\keyword{datasets}
