% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runspatialdecon.R
\name{runspatialdecon}
\alias{runspatialdecon}
\title{Run spatialdecon}
\usage{
runspatialdecon(object, ...)
}
\arguments{
\item{object}{An S4 object such as a Seurat object or a GeoMxSet object}

\item{...}{Arguments passed to spatialdecon}
}
\description{
Runs spatialdecon from an S4 object
}
